VIP-827-backend #time 3h #comment updated createNotification and fine-tuned other functions

+++ src/schemas/Responses.js
@@ -43,7 +43,7 @@ export const types = `
     enddate: Date!
     requested: Date!
     decided: Date
-    days: Int!
+    days: Float!
     status: VACATION_STATUS!
   }
